Fix exceptions showing to user with invalid configuration
When users incorrectly configure their credentials but still save the issue tracker binding, or maybe because they switch to a workspace without configuration data, the extension would fail loading in the background and show this via a nasty error popup.

These errors are now swallowed in code, which doesn't sound like best practice, but seems adequate in this case. Users should not be bothered with error popups if they are not actively using the extension. The exception message also doesn't contain any useful information about what happened from a user perspective.