Support empty branch prefix
With the previous implementation, it was impossible to not set a branch prefix in the preferences, because the code assumed that a null or empty string means, the user hasn't yet configured anything and so defaults should be used. By supplying a non-empty default, this resulted in users being forced to set a branch prefix or use said default.

Now the branch prefix field is empty by default and users can leave it like that if they chose to. This might also be a better default because it might be presumptuous to come up with a good one.