Implement GetTaskForBranch
This shows the task info associated with a branch in the branch explorer sidebar.

Unluckily, it wasn't straightforward to fetch the associated Codecks card from the full branch name. First, the full branch needs to be parsed to extract the task ID. In this case, this will produce the Codecks card ID label. However, if no branch prefix is used, it seems impossible to check if the resulting name is a valid ID or simply an untracked branch like 'main' other than trying the request.

To not make the request depend on the order of other requests which cached some of the request data (email-userid lookup), those requests were reimplemented as standalone calls for now. Later, it would probably be possible to cache the results in a more elaborate manner, but even better, use the power of the query API once the documentation has improved.