Reintroduce email cache to avoid rate-limiting errors
In a previous commit, I removed the user id to email cache to make the code less time dependent, but this now results in rate-limiting errors when requesting too many cards at once. As a quick fix, this commit used the cache again. In the future, this issue can likely be solved by using a more powerful query.