Improve internal error-handling and external error reporting
It turns out some exceptions are nicely reported in the Plastic user interface, e.g. anything thrown when trying to create a new branch from a task (GetPendingTasks).

Others method calls, however, open an error popup. Depending on the situation, some decisions about throwing or catching can be made. In most other calls it seems to make sense to catch exceptions and return empty collections instead. For example, when trying to create a new task in the task-on-changeset mode, LoadTasks is called. If this returns an empty collection, Plastic will already display a nice error message saying "No issue found with ID...".

Some methods like GetTaskForBranch are called very often (e.g. when selecting a branch in the branch explorer) and therefore they should not annoy the user with popups.