Overhaul code architecture
This is a really big refactoring (much bigger than we would want it to be when working together in a team).

The configuration values (user preferences) now have their own system which encapsulates configuration management. This moves most of the code from the extension factory into the new ConfigValues class.

Json queries are now no longer declared in C#, but instead embedded as json file resources in the assembly manifest. This makes maintaining these queries much easier as it is now possible to use any text editor or IDE with support for syntax checking, highlighting or automatic formatting. Much better than escaping and concatenating raw json strings in C#.

Error handling has been moved into the ExtensionErrorHandling class. It deals with the decision which errors can be reported to the user and which should be treated with silence.