Always keep logging enabled
Configuring the logging via the settings bool was producing inconsistent results. When the extension executes "for real", it respected the bool, but for example, when calling "Test Connection" the configuration returned by Plastic was sometimes null, resulting in the default value false being used for the logging.

I didn't understand this input, maybe it's a Plastic bug, but in any case it's too disruptive if the log enabled option is only respected most of the time. If it doesn't work everywhere, it can't be trusted, so I decided to always enable logging. It's not excessive anyway, so the performance cost should be minimal.