Merge logging and error handling to support better logs
For example, when the utils.dll is missing, there's an exception in the Connect method. However, we don't want to throw that, because it would show an error popup when opening the branches view in Plastic and prevent users from working at all. Later, when trying to create a new branch, a NullReference popup will show. The latter case is ok to signal that the extension is broken (after all the user can then disable it in the settings), but the info about what went wrong during connect is missing. Now, with the changes, we at least get a log that has some info in it: FileNotFound utils.dll during Connect.