Move queries from embedded resources into plain files
This will potentially allow users to modify/fix queries on their machines. There's no real reason to try and hide them. The API is documented online, there's nothing security critical in those files and the only reason I originally wanted to pack them was "to be neat and tidy". But by now, we have multiple files in the extension directory anyway, and it shouldn't be too hard for users, since they just take the zip and unzip it anyway.