Copy paste required source code from Codice utils.dll
We require the method to decrypt the user password from the configuration, but it's kind of annoying to have multiple dependencies shipped with the extension. In the future we should report this as a feature request to make the code available to extensions from the PlasticSCM host, it's likely an oversight on their part that it isn't.